requirements.txt
discord.py>=1.3.3
python-daemon>=2.2.4
python-dotenv>=0.13.0
